# Some minor performance improvements