**Before**
```
james@Jamess-Mac-mini ruby-liquid2 % bundle exec ruby performance/benchmark.rb 
ruby 3.4.3 (2025-04-14 revision d0b7e5b6a0) +PRISM [arm64-darwin24]
Warming up --------------------------------------
        parse (002):   293.000 i/100ms
       render (002):     1.113k i/100ms
         both (002):   219.000 i/100ms
Calculating -------------------------------------
        parse (002):      2.854k (± 0.5%) i/s  (350.38 μs/i) -     14.357k in   5.030505s
       render (002):     11.040k (± 0.6%) i/s   (90.58 μs/i) -     55.650k in   5.041144s
         both (002):      2.218k (± 1.0%) i/s  (450.90 μs/i) -     11.169k in   5.036595s
```

**After**
```
james@Jamess-Mac-mini ruby-liquid2 % bundle exec ruby performance/benchmark.rb 
ruby 3.4.3 (2025-04-14 revision d0b7e5b6a0) +PRISM [arm64-darwin24]
Warming up --------------------------------------
        parse (002):   312.000 i/100ms
       render (002):     1.160k i/100ms
         both (002):   232.000 i/100ms
Calculating -------------------------------------
        parse (002):      3.062k (± 1.2%) i/s  (326.57 μs/i) -     15.600k in   5.095298s
       render (002):     11.720k (± 1.2%) i/s   (85.32 μs/i) -     59.160k in   5.048413s
         both (002):      2.342k (± 1.3%) i/s  (427.00 μs/i) -     11.832k in   5.053072s
```

**Before (YJIT)**
```
james@Jamess-Mac-mini ruby-liquid2 % bundle exec ruby --yjit performance/benchmark.rb
ruby 3.4.3 (2025-04-14 revision d0b7e5b6a0) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
        parse (002):   476.000 i/100ms
       render (002):     2.755k i/100ms
         both (002):   401.000 i/100ms
Calculating -------------------------------------
        parse (002):      4.681k (± 1.3%) i/s  (213.64 μs/i) -     23.800k in   5.085559s
       render (002):     27.512k (± 0.9%) i/s   (36.35 μs/i) -    137.750k in   5.007297s
         both (002):      3.949k (± 1.3%) i/s  (253.22 μs/i) -     20.050k in   5.077941s
```

**After (YJIT)**
```
ruby 3.4.3 (2025-04-14 revision d0b7e5b6a0) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
        parse (002):   477.000 i/100ms
       render (002):     2.947k i/100ms
         both (002):   406.000 i/100ms
Calculating -------------------------------------
        parse (002):      4.755k (± 0.2%) i/s  (210.30 μs/i) -     23.850k in   5.015654s
       render (002):     29.242k (± 0.3%) i/s   (34.20 μs/i) -    147.350k in   5.038968s
         both (002):      4.051k (± 0.2%) i/s  (246.87 μs/i) -     20.300k in   5.011524s
```